training zoom

+++ save_zoomed_specs.py
@@ -3,6 +3,7 @@ import numpy as np
 from params import params
 import os
 from glob import glob
+from tqdm import tqdm
 
 
 def zoom_spec(spec, reduction_factor):
@@ -15,9 +16,11 @@ if __name__ == "__main__":
     spec_full_dir = os.path.join(params.spectrogram_dir_root, params.spectrogram_full_dir)
 
     spec_paths = glob(f"{spec_full_dir}/**/*.spec.npy", recursive=True)
-    for spec_path in spec_paths:
+    for spec_path in tqdm(spec_paths):
+        spec_reduced_path = os.path.join(spec_reduced_dir, os.path.relpath(spec_path, spec_full_dir))
+        if os.path.exists(spec_reduced_path): continue
         spec = np.load(spec_path)
         spec_reduced = zoom_spec(spec, reduction_factor)
-        spec_reduced_path = os.path.join(spec_reduced_dir, os.path.relpath(spec_path, spec_full_dir))
+
         os.makedirs(os.path.dirname(spec_reduced_path), exist_ok=True)
         np.save(spec_reduced_path, spec_reduced)
